docs: Add content to external-facing nada-dsl package README
We are starting to publish nada-dsl to TestPyPI and PyPI. The existing
nada-dsl package README was empty. This commit adds content to the
README so that the package, when published to PyPI, is a bit more
presentable and user-friendly. A brief description of the package and
references to links where users can find more info about Nada DSL were
added.