docs: Fix broken links in nada-dsl homepage published to PyPI
This commit introduces a new README reserved for reference by the
pyproject.toml file which includes fixes to broken links present in
PyPI.

The new README is added because the existing one now has details meant
for end-users AND us internally. To avoid confusion, the new README,
named with a .pyproject.md extensions, is reserved for end-users.