chore: Refer to renamed secrets in `build.yml` (#56)
This commit changes the reference to the app ID and private key values
used to generate app tokens used to close the nillion repo. This change
reflects the corresponding rename I performed on the GitHub App itself.
We have 2 registered GitHub Apps related to GitHub Actions use cases:

* Nillion Github Actions (already existed)
* And Nillion GitHub Actions Public (new as of today)

The latter is used for public repos and is currently limited to
nada-dsl.